feat(pubsublite): Move internal implementation details to internal/wire subpackage
